refactor(runtime): consolidate test files following Go naming convention
- Merged 13 separate test files into 3 properly named files:
  - runtime_test.go (tests runtime.go functions)
  - exit_test.go (tests exit.go functions)
  - system_test.go (tests system functions, already properly named)
- Removed files: comprehensive_test.go, coverage_test.go, error_path_test.go, missing_coverage_test.go, missing_lines_test.go, mock_error_test.go, panic_edge_cases_test.go, path_error_coverage_test.go, final_coverage_test.go, stack_edge_test.go, error_paths_coverage_test.go, exit_complete_test.go, exit_function_coverage_test.go
- Combined unit tests and benchmark tests in single files
- Maintains 100% test coverage and functionality
- Follows standard Go convention: {source_file}.go -> {source_file}_test.go
