--- runtime/comprehensive_test.go
@@ -1,264 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"testing"
-	"time"
-)
-
-// 最终综合测试，针对剩余的覆盖率缺口
-func TestComprehensiveFinalCoverage(t *testing.T) {
-	// 对于Exit函数，我们无法真实测试，因为它会终止进程
-	// 但我们可以测试它的所有组件和逻辑路径
-	t.Run("exit_function_component_test", func(t *testing.T) {
-		// 测试Exit函数中使用的所有组件
-
-		// 1. 测试os.FindProcess的成功路径
-		currentPID := os.Getpid()
-		process, err := os.FindProcess(currentPID)
-		if err != nil {
-			// 这对应Exit函数中的错误处理：log.Errorf("err:%v", err) + os.Exit(0)
-			t.Logf("Exit function error path: FindProcess failed: %v", err)
-		} else {
-			// 这对应Exit函数中的成功路径：log.Infof("will stop process:%d", process.Pid)
-			t.Logf("Exit function success path: found process %d", process.Pid)
-			
-			// 验证进程对象
-			if process.Pid != currentPID {
-				t.Errorf("Process PID mismatch: expected %d, got %d", currentPID, process.Pid)
-			}
-
-			// 注意：我们不能测试实际的Signal调用，因为它可能终止测试进程
-			// 但Exit函数的逻辑已经通过FindProcess测试覆盖了
-		}
-
-		// 2. 测试os.FindProcess的失败路径（模拟）
-		// 使用一个不太可能存在的PID
-		impossiblePID := -12345 // 负数PID通常无效
-		_, err = os.FindProcess(impossiblePID)
-		if err != nil {
-			t.Logf("Exit function error simulation: FindProcess with invalid PID failed: %v", err)
-		}
-
-		t.Log("Exit function component testing completed")
-	})
-
-	// WaitExit函数测试：我们不能让它真正阻塞，但可以测试其组件
-	t.Run("wait_exit_component_test", func(t *testing.T) {
-		// WaitExit的实现是：
-		// sign := GetExitSign()
-		// <-sign
-		// 我们已经测试了GetExitSign，WaitExit只是从通道读取
-
-		sigCh := GetExitSign()
-		if sigCh == nil {
-			t.Fatal("GetExitSign should not return nil")
-		}
-
-		// 测试通道的读取机制（模拟WaitExit的核心逻辑）
-		done := make(chan bool, 1)
-		go func() {
-			// 模拟WaitExit的行为，但有超时
-			select {
-			case sig := <-sigCh:
-				t.Logf("WaitExit simulation: received signal %v", sig)
-				done <- true
-			case <-time.After(10 * time.Millisecond):
-				t.Log("WaitExit simulation: timeout (expected in test)")
-				done <- false
-			}
-		}()
-
-		// 等待完成
-		<-done
-		
-		t.Log("WaitExit component testing completed")
-	})
-
-	// 测试PrintStack的所有可能分支
-	t.Run("print_stack_complete_branch_test", func(t *testing.T) {
-		// PrintStack的逻辑：
-		// st := debug.Stack()
-		// if len(st) > 0 {
-		//     log.Error("dump stack:")
-		//     log.Error(string(st))
-		// } else {
-		//     log.Error("stack is empty")
-		// }
-
-		// 在正常情况下，debug.Stack()总是返回非空数据
-		// 所以我们主要覆盖有栈信息的分支
-		PrintStack()
-
-		// 虽然我们无法真正触发空栈情况，但我们已经覆盖了函数的主要逻辑
-		t.Log("PrintStack all reachable branches tested")
-	})
-
-	// 测试CachePanicWithHandle的完整逻辑，特别是栈处理分支
-	t.Run("cache_panic_handle_complete_stack_test", func(t *testing.T) {
-		// CachePanicWithHandle的关键逻辑包括栈处理：
-		// st := debug.Stack()
-		// if len(st) > 0 {
-		//     log.Errorf("dump stack (%s):", err)
-		//     lines := strings.Split(string(st), "\n")
-		//     for _, line := range lines {
-		//         log.Error("  ", line)
-		//     }
-		// } else {
-		//     log.Errorf("stack is empty (%s)", err)
-		// }
-
-		var capturedErr interface{}
-		func() {
-			defer func() {
-				recover() // 防止panic传播到测试框架
-			}()
-			defer CachePanicWithHandle(func(err interface{}) {
-				capturedErr = err
-			})
-			panic("comprehensive stack test")
-		}()
-
-		if capturedErr == nil {
-			t.Error("Expected panic to be captured")
-		}
-
-		t.Log("CachePanicWithHandle complete stack testing finished")
-	})
-
-	// 尝试触发路径函数的错误分支（虽然在正常环境中很难）
-	t.Run("path_functions_error_branch_attempt", func(t *testing.T) {
-		// 这些函数的错误分支在正常环境中很难触发：
-		// - ExecDir: return "" if os.Executable() fails
-		// - ExecFile: return "" if os.Executable() fails  
-		// - Pwd: return "" if os.Getwd() fails
-
-		// 我们只能测试正常情况，并记录结果
-		functions := map[string]func() string{
-			"ExecDir":  ExecDir,
-			"ExecFile": ExecFile,  
-			"Pwd":      Pwd,
-		}
-
-		for name, fn := range functions {
-			result := fn()
-			if result == "" {
-				t.Logf("%s error branch triggered (returned empty)", name)
-			} else {
-				t.Logf("%s normal path: %s", name, result)
-			}
-		}
-
-		// 注意：在正常的测试环境中，os.Executable()和os.Getwd()几乎总是成功的
-		// 要触发错误分支需要非常特殊的环境条件，这在单元测试中不现实
-		t.Log("Path functions error branch testing attempted")
-	})
-}
-
-// 额外的边界情况测试
-func TestEdgeCasesForMaxCoverage(t *testing.T) {
-	t.Run("multiple_panic_handle_calls", func(t *testing.T) {
-		// 测试连续的panic处理调用
-		for i := 0; i < 3; i++ {
-			func() {
-				defer func() {
-					recover()
-				}()
-				defer CachePanicWithHandle(func(err interface{}) {
-					t.Logf("Handle call %d: %v", i, err)
-				})
-				panic("edge case " + string(rune('A'+i)))
-			}()
-		}
-		t.Log("Multiple panic handle calls completed")
-	})
-
-	t.Run("concurrent_function_calls", func(t *testing.T) {
-		// 并发调用所有函数以确保线程安全和覆盖率
-		const numGoroutines = 3
-		done := make(chan bool, numGoroutines)
-
-		for i := 0; i < numGoroutines; i++ {
-			go func(id int) {
-				defer func() {
-					done <- true
-				}()
-
-				// 调用所有主要函数
-				_ = IsWindows()
-				_ = IsDarwin()
-				_ = IsLinux()
-				_ = ExecDir()
-				_ = ExecFile()
-				_ = Pwd()
-				_ = UserHomeDir()
-				_ = UserConfigDir()
-				_ = UserCacheDir()
-				_ = LazyConfigDir()
-				_ = LazyCacheDir()
-				_ = GetExitSign()
-				PrintStack()
-				
-				func() {
-					defer CachePanic()
-				}()
-
-				t.Logf("Goroutine %d completed all function calls", id)
-			}(i)
-		}
-
-		// 等待所有goroutine完成
-		for i := 0; i < numGoroutines; i++ {
-			<-done
-		}
-
-		t.Log("Concurrent function calls completed")
-	})
-
-	t.Run("stress_test_all_functions", func(t *testing.T) {
-		// 压力测试：多次调用所有函数
-		const iterations = 10
-
-		for i := 0; i < iterations; i++ {
-			// 系统检测函数
-			_ = IsWindows()
-			_ = IsDarwin()  
-			_ = IsLinux()
-
-			// 路径函数
-			_ = ExecDir()
-			_ = ExecFile()
-			_ = Pwd()
-
-			// 用户目录函数
-			_ = UserHomeDir()
-			_ = UserConfigDir()
-			_ = UserCacheDir()
-			_ = LazyConfigDir()
-			_ = LazyCacheDir()
-
-			// 信号函数
-			sigCh := GetExitSign()
-			if sigCh == nil {
-				t.Error("GetExitSign should not return nil")
-			}
-
-			// 堆栈和panic函数
-			PrintStack()
-			
-			func() {
-				defer CachePanic()
-			}()
-
-			func() {
-				defer CachePanicWithHandle(nil)
-			}()
-
-			if i%5 == 0 {
-				t.Logf("Stress test iteration %d completed", i)
-			}
-		}
-
-		t.Log("Stress test completed")
-	})
-}
