--- runtime/coverage_test.go
@@ -1,397 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"testing"
-)
-
-// 测试错误分支以提高覆盖率
-func TestErrorBranches(t *testing.T) {
-	t.Run("test_exec_dir_error_handling", func(t *testing.T) {
-		// ExecDir函数的错误分支很难触发，因为os.Executable()通常都会成功
-		// 我们主要验证函数能正常执行
-		result := ExecDir()
-		if result == "" {
-			t.Log("ExecDir returned empty string (may be due to error)")
-		} else {
-			t.Logf("ExecDir returned: %s", result)
-		}
-	})
-
-	t.Run("test_exec_file_error_handling", func(t *testing.T) {
-		// ExecFile函数的错误分支也很难触发
-		result := ExecFile()
-		if result == "" {
-			t.Log("ExecFile returned empty string (may be due to error)")
-		} else {
-			t.Logf("ExecFile returned: %s", result)
-		}
-	})
-
-	t.Run("test_pwd_error_handling", func(t *testing.T) {
-		// Pwd函数的错误分支可能在某些特殊情况下触发
-		result := Pwd()
-		if result == "" {
-			t.Log("Pwd returned empty string (may be due to error)")
-		} else {
-			t.Logf("Pwd returned: %s", result)
-		}
-	})
-}
-
-// 测试PrintStack的不同情况
-func TestPrintStackCoverage(t *testing.T) {
-	t.Run("print_stack_with_valid_stack", func(t *testing.T) {
-		// 测试PrintStack在有有效堆栈时的行为
-		PrintStack()
-		t.Log("PrintStack executed with valid stack")
-	})
-}
-
-// 测试panic处理的边界情况
-func TestPanicHandlingEdgeCases(t *testing.T) {
-	t.Run("cache_panic_with_handle_edge_cases", func(t *testing.T) {
-		// 测试handle函数被正确调用的边界情况
-
-		handleCallCount := 0
-		handle := func(err interface{}) {
-			handleCallCount++
-			// 验证错误内容
-			if err == nil {
-				t.Error("Handle should receive non-nil error")
-			}
-		}
-
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Test recovered: %v", r)
-			}
-			if handleCallCount != 1 {
-				t.Errorf("Handle should be called exactly once, was called %d times", handleCallCount)
-			}
-		}()
-
-		// 触发panic并验证handle被调用
-		func() {
-			defer CachePanicWithHandle(handle)
-			panic("edge case panic")
-		}()
-	})
-
-	t.Run("cache_panic_with_handle_nil_error", func(t *testing.T) {
-		// 测试当没有panic时，CachePanicWithHandle的行为
-		handleCalled := false
-		handle := func(err interface{}) {
-			handleCalled = true
-		}
-
-		func() {
-			defer CachePanicWithHandle(handle)
-			// 不产生panic
-		}()
-
-		if handleCalled {
-			t.Error("Handle should not be called when no panic occurs")
-		}
-		t.Log("CachePanicWithHandle correctly handled no-panic case")
-	})
-}
-
-// 测试系统函数的边界条件
-func TestSystemFunctionsCoverage(t *testing.T) {
-	t.Run("test_all_system_detection_functions", func(t *testing.T) {
-		// 确保所有系统检测函数都被调用以提高覆盖率
-		windows := IsWindows()
-		darwin := IsDarwin()
-		linux := IsLinux()
-
-		t.Logf("System detection: Windows=%v, Darwin=%v, Linux=%v", windows, darwin, linux)
-
-		// 验证只有一个为true
-		trueCount := 0
-		if windows {
-			trueCount++
-		}
-		if darwin {
-			trueCount++
-		}
-		if linux {
-			trueCount++
-		}
-
-		if trueCount != 1 {
-			t.Errorf("Expected exactly one OS to be detected, got %d", trueCount)
-		}
-	})
-}
-
-// 测试目录函数的各种情况
-func TestDirectoryFunctionsCoverage(t *testing.T) {
-	t.Run("test_user_directory_functions", func(t *testing.T) {
-		// 测试所有用户目录函数
-		home := UserHomeDir()
-		config := UserConfigDir()
-		cache := UserCacheDir()
-		lazyConfig := LazyConfigDir()
-		lazyCache := LazyCacheDir()
-
-		t.Logf("UserHomeDir: %s", home)
-		t.Logf("UserConfigDir: %s", config)
-		t.Logf("UserCacheDir: %s", cache)
-		t.Logf("LazyConfigDir: %s", lazyConfig)
-		t.Logf("LazyCacheDir: %s", lazyCache)
-
-		// 基本有效性检查
-		directories := []string{home, config, cache, lazyConfig, lazyCache}
-		for i, dir := range directories {
-			if dir != "" && dir[0] != '/' && dir[0] != '\\' {
-				t.Logf("Directory %d may not be absolute: %s", i, dir)
-			}
-		}
-	})
-}
-
-// 测试路径操作函数
-func TestPathFunctionsCoverage(t *testing.T) {
-	t.Run("test_exec_and_pwd_functions", func(t *testing.T) {
-		execDir := ExecDir()
-		execFile := ExecFile()
-		pwd := Pwd()
-
-		t.Logf("ExecDir: %s", execDir)
-		t.Logf("ExecFile: %s", execFile)
-		t.Logf("Pwd: %s", pwd)
-
-		// 验证基本属性
-		if execDir != "" && execFile != "" {
-			// execFile应该在execDir中
-			if len(execFile) <= len(execDir) {
-				t.Log("ExecFile path may not contain ExecDir")
-			}
-		}
-
-		if pwd == "" {
-			t.Error("Pwd should not return empty string under normal conditions")
-		}
-	})
-}
-
-// 极端情况测试
-func TestExtremeCases(t *testing.T) {
-	t.Run("test_functions_under_restricted_environment", func(t *testing.T) {
-		// 在受限环境下测试函数行为
-		// 这些测试主要是为了触发错误处理分支
-
-		// 备份当前环境变量
-		originalHome := os.Getenv("HOME")
-
-		// 临时清空HOME环境变量
-		os.Setenv("HOME", "")
-
-		// 测试在受限环境下的行为
-		home := UserHomeDir()
-		t.Logf("UserHomeDir with empty HOME: %s", home)
-
-		// 恢复环境变量
-		os.Setenv("HOME", originalHome)
-	})
-}
-
-// 测试Exit函数的完整逻辑但不实际退出
-func TestExitFunctionLogic(t *testing.T) {
-	t.Run("test_exit_function_process_finding", func(t *testing.T) {
-		// 测试Exit函数中的进程查找逻辑
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Errorf("Exit function should be able to find current process: %v", err)
-			return
-		}
-
-		if process.Pid != os.Getpid() {
-			t.Errorf("Process PID should match current PID: expected %d, got %d", os.Getpid(), process.Pid)
-		}
-
-		// 验证进程对象是有效的
-		if process == nil {
-			t.Error("Process should not be nil")
-		}
-
-		t.Log("Exit function components (process operations) work correctly")
-	})
-
-	t.Run("test_exit_function_invalid_process", func(t *testing.T) {
-		// 测试FindProcess对无效PID的处理
-		// 使用一个非常大的PID，通常不会存在
-		invalidPID := 999999999
-		process, err := os.FindProcess(invalidPID)
-		
-		// 在某些系统上，FindProcess可能不会立即返回错误
-		// 但后续的Signal操作会失败
-		if err == nil && process != nil {
-			// 尝试发送信号来测试进程是否真的存在
-			signalErr := process.Signal(os.Interrupt)
-			if signalErr != nil {
-				t.Logf("Signal to invalid process failed as expected: %v", signalErr)
-			} else {
-				t.Logf("Signal to process %d succeeded (process may exist)", invalidPID)
-			}
-		} else {
-			t.Logf("FindProcess returned error for invalid PID as expected: %v", err)
-		}
-	})
-
-	t.Run("test_exit_signal_mechanism", func(t *testing.T) {
-		// 测试Exit函数中使用的信号机制
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Fatalf("Should be able to find current process: %v", err)
-		}
-
-		// 测试能否发送信号（但使用非致命信号进行测试）
-		// 注意：我们不能测试实际的Exit函数，因为它会终止进程
-		// 但我们可以验证Signal方法的工作原理
-		err = process.Signal(os.Interrupt)
-		if err != nil {
-			t.Logf("Signal sending failed (may be expected on some platforms): %v", err)
-		} else {
-			t.Log("Signal mechanism works correctly")
-		}
-	})
-}
-
-// 验证修复后的逻辑
-func TestFixedLogic(t *testing.T) {
-	t.Run("test_user_cache_dir_variable_name_fix", func(t *testing.T) {
-		// 验证UserCacheDir函数使用正确的变量名
-		result := UserCacheDir()
-
-		// 这个函数之前有bug：变量名错误（execPath而不是path）
-		// 现在应该工作正常
-		if result == "" {
-			t.Log("UserCacheDir returned empty (may be normal in some environments)")
-		} else {
-			t.Logf("UserCacheDir correctly returned: %s", result)
-		}
-	})
-
-	t.Run("test_exit_function_components", func(t *testing.T) {
-		// 验证Exit函数使用的组件（不发送信号）
-
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Errorf("Should be able to find current process: %v", err)
-		}
-
-		if process.Pid != os.Getpid() {
-			t.Errorf("Process PID mismatch: expected %d, got %d", os.Getpid(), process.Pid)
-		}
-
-		t.Log("Exit function components (process finding) work correctly")
-	})
-}
-
-// 测试PrintStack函数的完整覆盖率
-func TestPrintStackEmptyCase(t *testing.T) {
-	t.Run("test_print_stack_normal_case", func(t *testing.T) {
-		// 测试PrintStack函数在正常情况下的行为
-		// 在正常情况下，debug.Stack()应该返回非空的堆栈信息
-		PrintStack()
-		t.Log("PrintStack executed successfully with stack trace")
-	})
-
-	t.Run("test_print_stack_logging", func(t *testing.T) {
-		// 确保PrintStack函数的所有分支都被执行
-		// 这个测试主要是确保覆盖PrintStack函数中的所有代码路径
-		defer func() {
-			if r := recover(); r != nil {
-				t.Errorf("PrintStack should not panic: %v", r)
-			}
-		}()
-
-		PrintStack()
-		t.Log("PrintStack function completed without issues")
-	})
-}
-
-// 测试CachePanicWithHandle的完整覆盖率
-func TestCachePanicCompleteCase(t *testing.T) {
-	t.Run("test_cache_panic_with_handle_stack_logging", func(t *testing.T) {
-		// 测试CachePanicWithHandle中堆栈记录的完整逻辑
-		handleCalled := false
-		var capturedError interface{}
-
-		handle := func(err interface{}) {
-			handleCalled = true
-			capturedError = err
-		}
-
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Outer recover caught: %v", r)
-			}
-
-			if !handleCalled {
-				t.Error("Handle function should have been called")
-			}
-
-			if capturedError == nil {
-				t.Error("Handle should have received error")
-			}
-		}()
-
-		// 触发panic以测试CachePanicWithHandle的完整逻辑
-		func() {
-			defer CachePanicWithHandle(handle)
-			panic("test panic for complete coverage")
-		}()
-	})
-
-	t.Run("test_cache_panic_no_handle", func(t *testing.T) {
-		// 测试CachePanicWithHandle在handle为nil时的行为
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Panic with nil handle recovered: %v", r)
-			}
-		}()
-
-		func() {
-			defer CachePanicWithHandle(nil)
-			panic("test panic with nil handle for coverage")
-		}()
-	})
-}
-
-// 测试错误路径的完整覆盖率
-func TestErrorPathsCoverage(t *testing.T) {
-	t.Run("test_exec_dir_error_path", func(t *testing.T) {
-		// 尽力测试ExecDir的错误路径
-		// 在正常情况下，os.Executable()很少失败，但我们确保函数被完全测试
-		result := ExecDir()
-		// 如果返回空字符串，说明进入了错误处理分支
-		if result == "" {
-			t.Log("ExecDir error path triggered (os.Executable failed)")
-		} else {
-			t.Logf("ExecDir success path: %s", result)
-		}
-	})
-
-	t.Run("test_exec_file_error_path", func(t *testing.T) {
-		// 尽力测试ExecFile的错误路径
-		result := ExecFile()
-		if result == "" {
-			t.Log("ExecFile error path triggered (os.Executable failed)")
-		} else {
-			t.Logf("ExecFile success path: %s", result)
-		}
-	})
-
-	t.Run("test_pwd_error_path", func(t *testing.T) {
-		// 尽力测试Pwd的错误路径
-		result := Pwd()
-		if result == "" {
-			t.Log("Pwd error path triggered (os.Getwd failed)")
-		} else {
-			t.Logf("Pwd success path: %s", result)
-		}
-	})
-}
