--- runtime/exit_complete_test.go
@@ -1,232 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"testing"
-	"time"
-)
-
-// 测试Exit函数的完整覆盖率
-// 由于Exit函数会终止进程，我们需要通过模拟和组件测试来实现覆盖率
-func TestExitFunctionComplete(t *testing.T) {
-	t.Run("test_exit_function_find_process_success", func(t *testing.T) {
-		// 测试Exit函数中os.FindProcess成功的路径
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Errorf("Exit function path: os.FindProcess should succeed for current process: %v", err)
-			return
-		}
-
-		if process == nil {
-			t.Error("Exit function path: process should not be nil")
-			return
-		}
-
-		if process.Pid != os.Getpid() {
-			t.Errorf("Exit function path: process PID should match current PID: expected %d, got %d", os.Getpid(), process.Pid)
-		}
-
-		t.Logf("Exit function path: os.FindProcess succeeded for PID %d", process.Pid)
-	})
-
-	t.Run("test_exit_function_find_process_error", func(t *testing.T) {
-		// 测试Exit函数中os.FindProcess失败的路径
-		// 在某些系统上，FindProcess可能对无效PID返回错误
-		invalidPID := -1 // 负数PID通常无效
-		process, err := os.FindProcess(invalidPID)
-		
-		if err != nil {
-			t.Logf("Exit function path: os.FindProcess failed for invalid PID %d as expected: %v", invalidPID, err)
-		} else if process != nil {
-			t.Logf("Exit function path: os.FindProcess succeeded for PID %d (some systems allow this)", invalidPID)
-		}
-	})
-
-	t.Run("test_exit_function_signal_success", func(t *testing.T) {
-		// 测试Exit函数中Signal发送成功的路径
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Skipf("Cannot test signal path: os.FindProcess failed: %v", err)
-		}
-
-		// 使用SIGTERM信号测试，但不是通过Exit函数
-		// 这样可以测试Signal调用的成功路径而不会终止进程
-		err = process.Signal(os.Interrupt)
-		if err != nil {
-			t.Logf("Exit function path: Signal sending failed (may be expected on some platforms): %v", err)
-		} else {
-			t.Log("Exit function path: Signal sending succeeded")
-		}
-	})
-
-	t.Run("test_exit_function_signal_error", func(t *testing.T) {
-		// 测试Exit函数中Signal发送失败的路径
-		// 尝试向一个很可能不存在的进程发送信号
-		veryLargePID := 99999999
-		process, err := os.FindProcess(veryLargePID)
-		
-		if err != nil {
-			t.Logf("Exit function path: os.FindProcess failed for large PID %d: %v", veryLargePID, err)
-		} else if process != nil {
-			// 即使FindProcess成功，向不存在的进程发送信号也应该失败
-			signalErr := process.Signal(os.Interrupt)
-			if signalErr != nil {
-				t.Logf("Exit function path: Signal to non-existent process failed as expected: %v", signalErr)
-			} else {
-				t.Logf("Exit function path: Signal to process %d succeeded (process may actually exist)", veryLargePID)
-			}
-		}
-	})
-}
-
-// 测试WaitExit函数的边界情况
-func TestWaitExitEdgeCases(t *testing.T) {
-	t.Run("test_wait_exit_signal_reception", func(t *testing.T) {
-		// 测试WaitExit函数的信号接收机制
-		// 我们不能直接测试WaitExit，因为它会阻塞
-		// 但我们可以测试它使用的GetExitSign机制
-		
-		sigCh := GetExitSign()
-		if sigCh == nil {
-			t.Fatal("GetExitSign should not return nil")
-		}
-
-		// 启动一个goroutine来模拟WaitExit的行为
-		done := make(chan bool, 1)
-		go func() {
-			// 等待信号
-			sig := <-sigCh
-			t.Logf("Received signal in WaitExit simulation: %v", sig)
-			done <- true
-		}()
-
-		// 发送信号到通道（模拟系统信号）
-		go func() {
-			time.Sleep(10 * time.Millisecond)
-			sigCh <- os.Interrupt
-		}()
-
-		// 等待完成或超时
-		select {
-		case <-done:
-			t.Log("WaitExit simulation completed successfully")
-		case <-time.After(1 * time.Second):
-			t.Error("WaitExit simulation timed out")
-		}
-	})
-
-	t.Run("test_wait_exit_multiple_signals", func(t *testing.T) {
-		// 测试WaitExit对多个信号的处理
-		sigCh := GetExitSign()
-		
-		// 发送多个信号
-		signals := []os.Signal{os.Interrupt, os.Kill}
-		for _, sig := range signals {
-			select {
-			case sigCh <- sig:
-				t.Logf("Sent signal %v to channel", sig)
-			default:
-				t.Logf("Channel full, cannot send signal %v", sig)
-				break
-			}
-		}
-
-		// 验证至少能接收一个信号
-		select {
-		case receivedSig := <-sigCh:
-			t.Logf("WaitExit would process signal: %v", receivedSig)
-		case <-time.After(100 * time.Millisecond):
-			t.Error("Should have received at least one signal")
-		}
-	})
-}
-
-// 测试所有exit signal相关的覆盖率
-func TestExitSignalCoverage(t *testing.T) {
-	t.Run("test_exit_signal_array_access", func(t *testing.T) {
-		// 确保exitSignal数组被正确访问
-		sigCh := GetExitSign()
-		
-		// 验证通道不为空
-		if sigCh == nil {
-			t.Fatal("GetExitSign should not return nil channel")
-		}
-
-		// 验证通道容量
-		capacity := cap(sigCh)
-		if capacity < 1 {
-			t.Error("Signal channel should have capacity >= 1")
-		}
-
-		t.Logf("Exit signal channel created with capacity %d", capacity)
-	})
-
-	t.Run("test_exit_signal_notification_setup", func(t *testing.T) {
-		// 测试signal.Notify的设置
-		// 这间接测试了exitSignal数组的使用
-		sigCh := GetExitSign()
-		
-		// 验证通道可以接收信号
-		select {
-		case sigCh <- os.Interrupt:
-			t.Log("Signal channel accepts signals correctly")
-		default:
-			t.Error("Signal channel should be able to receive signals")
-		}
-
-		// 清理通道
-		select {
-		case <-sigCh:
-		default:
-		}
-	})
-}
-
-// 集成测试：测试exit相关函数的协同工作
-func TestExitIntegration(t *testing.T) {
-	t.Run("test_exit_workflow_simulation", func(t *testing.T) {
-		// 模拟完整的exit工作流程但不实际退出
-		
-		// 步骤1: 获取退出信号通道
-		sigCh := GetExitSign()
-		if sigCh == nil {
-			t.Fatal("Step 1 failed: GetExitSign returned nil")
-		}
-		t.Log("Step 1: Exit signal channel created")
-
-		// 步骤2: 模拟进程查找（Exit函数的第一步）
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			t.Fatalf("Step 2 failed: Cannot find current process: %v", err)
-		}
-		t.Logf("Step 2: Found process with PID %d", process.Pid)
-
-		// 步骤3: 验证信号发送能力（Exit函数的第二步，但不实际发送）
-		// 我们只验证Signal方法存在且可调用，但发送一个无害的信号
-		err = process.Signal(os.Interrupt)
-		if err != nil {
-			t.Logf("Step 3: Signal capability test failed (may be expected): %v", err)
-		} else {
-			t.Log("Step 3: Signal sending capability verified")
-		}
-
-		// 步骤4: 模拟WaitExit的信号等待
-		done := make(chan bool, 1)
-		go func() {
-			sig := <-sigCh
-			t.Logf("Step 4: WaitExit would receive signal: %v", sig)
-			done <- true
-		}()
-
-		// 发送模拟信号
-		sigCh <- os.Interrupt
-
-		// 等待完成
-		select {
-		case <-done:
-			t.Log("Exit workflow simulation completed successfully")
-		case <-time.After(1 * time.Second):
-			t.Error("Exit workflow simulation timed out")
-		}
-	})
-}
