--- runtime/final_coverage_test.go
@@ -1,216 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"testing"
-	"time"
-)
-
-// 专门为提高覆盖率而设计的测试
-func TestFinalCoverage(t *testing.T) {
-	t.Run("test_exit_function_without_actual_exit", func(t *testing.T) {
-		// 测试Exit函数的组件但不实际调用Exit
-		// 这样可以覆盖Exit函数中的代码路径
-		
-		// 测试os.FindProcess成功的情况
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			// 这里覆盖了Exit函数中err != nil的分支
-			t.Logf("Exit function error path: os.FindProcess failed: %v", err)
-		} else {
-			// 这里覆盖了Exit函数中process不为nil的分支
-			t.Logf("Exit function success path: found process PID %d", process.Pid)
-			
-			// 注意：我们不能真的调用process.Signal来终止进程
-			// 但我们可以测试Signal方法的存在性
-			if process.Pid == os.Getpid() {
-				t.Log("Exit function would signal current process")
-			}
-		}
-	})
-
-	t.Run("test_print_stack_all_branches", func(t *testing.T) {
-		// 确保PrintStack函数的所有代码分支都被覆盖
-		// 在正常情况下，debug.Stack()总是返回非空数据
-		PrintStack()
-		t.Log("PrintStack function executed")
-	})
-
-	t.Run("test_cache_panic_handle_all_branches", func(t *testing.T) {
-		// 测试CachePanicWithHandle的所有代码路径
-		
-		// 路径1: 没有panic发生
-		func() {
-			defer CachePanicWithHandle(nil)
-			// 正常执行，不panic
-		}()
-		
-		// 路径2: 有panic，handle为nil
-		func() {
-			defer func() {
-				recover() // 静默恢复，避免影响测试
-			}()
-			defer CachePanicWithHandle(nil)
-			panic("test nil handle")
-		}()
-		
-		// 路径3: 有panic，handle不为nil
-		handleCalled := false
-		func() {
-			defer func() {
-				recover() // 静默恢复，避免影响测试
-			}()
-			defer CachePanicWithHandle(func(err interface{}) {
-				handleCalled = true
-			})
-			panic("test with handle")
-		}()
-		
-		if !handleCalled {
-			t.Error("Handle should have been called")
-		}
-	})
-
-	t.Run("test_path_functions_error_branches", func(t *testing.T) {
-		// 测试路径函数的错误分支
-		// 这些函数在正常情况下不会出错，但我们确保它们被调用
-		
-		execDir := ExecDir()
-		execFile := ExecFile()
-		pwd := Pwd()
-		
-		// 记录结果以确保函数被执行
-		results := []string{execDir, execFile, pwd}
-		for i, result := range results {
-			if result == "" {
-				t.Logf("Path function %d returned empty (error branch)", i)
-			} else {
-				t.Logf("Path function %d returned: %s", i, result)
-			}
-		}
-	})
-
-	t.Run("test_user_directory_functions_complete", func(t *testing.T) {
-		// 确保所有用户目录函数都被完全执行
-		functions := map[string]func() string{
-			"UserHomeDir":   UserHomeDir,
-			"UserConfigDir": UserConfigDir,
-			"UserCacheDir":  UserCacheDir,
-			"LazyConfigDir": LazyConfigDir,
-			"LazyCacheDir":  LazyCacheDir,
-		}
-		
-		for name, fn := range functions {
-			result := fn()
-			t.Logf("%s: %s", name, result)
-		}
-	})
-
-	t.Run("test_system_detection_complete", func(t *testing.T) {
-		// 确保所有系统检测函数都被执行
-		windows := IsWindows()
-		darwin := IsDarwin()
-		linux := IsLinux()
-		
-		t.Logf("System detection - Windows: %v, Darwin: %v, Linux: %v", windows, darwin, linux)
-		
-		// 验证只有一个为true
-		count := 0
-		if windows {
-			count++
-		}
-		if darwin {
-			count++
-		}
-		if linux {
-			count++
-		}
-		
-		if count != 1 {
-			t.Errorf("Expected exactly one OS detection to be true, got %d", count)
-		}
-	})
-
-	t.Run("test_signal_channel_without_interference", func(t *testing.T) {
-		// 测试信号通道但不发送会干扰测试的信号
-		sigCh := GetExitSign()
-		if sigCh == nil {
-			t.Fatal("GetExitSign should not return nil")
-		}
-		
-		// 验证通道属性但不发送实际信号
-		select {
-		case <-sigCh:
-			t.Log("Signal channel had pending signal")
-		case <-time.After(1 * time.Millisecond):
-			t.Log("Signal channel is empty as expected")
-		}
-	})
-}
-
-// 专门测试CachePanic函数
-func TestCachePanicComplete(t *testing.T) {
-	t.Run("test_cache_panic_no_panic", func(t *testing.T) {
-		// 测试CachePanic在没有panic时的行为
-		func() {
-			defer CachePanic()
-			// 正常执行
-		}()
-		t.Log("CachePanic handled no-panic case")
-	})
-
-	t.Run("test_cache_panic_with_panic", func(t *testing.T) {
-		// 测试CachePanic在有panic时的行为
-		func() {
-			defer func() {
-				recover() // 静默恢复，避免影响测试
-			}()
-			defer CachePanic()
-			panic("test cache panic")
-		}()
-		t.Log("CachePanic handled panic case")
-	})
-}
-
-// 确保所有函数都被测试到
-func TestAllFunctionsCalled(t *testing.T) {
-	t.Run("comprehensive_function_coverage", func(t *testing.T) {
-		// 系统检测
-		_ = IsWindows()
-		_ = IsDarwin()
-		_ = IsLinux()
-		
-		// 路径函数
-		_ = ExecDir()
-		_ = ExecFile()
-		_ = Pwd()
-		
-		// 用户目录函数
-		_ = UserHomeDir()
-		_ = UserConfigDir()
-		_ = UserCacheDir()
-		_ = LazyConfigDir()
-		_ = LazyCacheDir()
-		
-		// 信号函数
-		_ = GetExitSign()
-		
-		// 堆栈和panic函数
-		PrintStack()
-		
-		// 测试CachePanic和CachePanicWithHandle的各种情况
-		func() {
-			defer CachePanic()
-		}()
-		
-		func() {
-			defer CachePanicWithHandle(nil)
-		}()
-		
-		func() {
-			defer CachePanicWithHandle(func(interface{}) {})
-		}()
-		
-		t.Log("All functions have been called for coverage")
-	})
-}
