--- runtime/missing_coverage_test.go
@@ -1,265 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"runtime/debug"
-	"testing"
-)
-
-// 专门针对缺失的覆盖率行进行测试
-func TestMissingCoverage(t *testing.T) {
-	// 测试Exit函数的逻辑，但不真的退出
-	t.Run("exit_function_logic_simulation", func(t *testing.T) {
-		// 模拟Exit函数的关键逻辑路径
-
-		// 路径1: os.FindProcess成功的情况
-		process, err := os.FindProcess(os.Getpid())
-		if err != nil {
-			// 这对应Exit函数中log.Errorf("err:%v", err)和os.Exit(0)的路径
-			t.Logf("Exit simulation: FindProcess failed: %v", err)
-		} else {
-			// 这对应Exit函数中log.Infof("will stop process:%d", process.Pid)的路径
-			t.Logf("Exit simulation: Found process %d", process.Pid)
-
-			// 注意: 我们不调用实际的Signal因为会终止测试进程
-			// 但我们可以验证Signal方法的存在和可调用性
-			// 这模拟了Exit函数中err = process.Signal(os.Interrupt)的路径
-			
-			// 模拟signal失败的情况（不实际发送信号）
-			// 这对应Exit函数中Signal失败后的错误处理路径
-			t.Log("Exit simulation: Would send signal to process")
-		}
-	})
-
-	// 测试PrintStack函数的空堆栈情况（理论上不可能，但为了覆盖率）
-	t.Run("print_stack_edge_case", func(t *testing.T) {
-		// 测试PrintStack函数，确保所有分支都被覆盖
-		// 在正常情况下debug.Stack()总是返回非空数据，所以我们主要确保函数被调用
-
-		// 直接调用PrintStack以覆盖所有代码路径
-		PrintStack()
-
-		// 验证debug.Stack()的行为以理解PrintStack的逻辑
-		stack := debug.Stack()
-		if len(stack) == 0 {
-			// 这种情况极其罕见，但如果发生会对应PrintStack中的else分支
-			t.Log("PrintStack edge case: empty stack detected")
-		} else {
-			// 这是正常情况，对应PrintStack中的if len(st) > 0分支
-			t.Logf("PrintStack normal case: stack size %d bytes", len(stack))
-		}
-	})
-
-	// 测试CachePanicWithHandle的空堆栈路径（理论情况）
-	t.Run("cache_panic_stack_edge_case", func(t *testing.T) {
-		// 测试CachePanicWithHandle中关于堆栈处理的所有分支
-		handleCalled := false
-		var capturedError interface{}
-
-		handle := func(err interface{}) {
-			handleCalled = true
-			capturedError = err
-		}
-
-		func() {
-			defer func() {
-				recover() // 防止panic传播
-			}()
-			defer CachePanicWithHandle(handle)
-			
-			// 触发panic以测试堆栈处理逻辑
-			panic("test for stack handling")
-		}()
-
-		if !handleCalled {
-			t.Error("Handle should have been called")
-		}
-		if capturedError == nil {
-			t.Error("Error should have been captured")
-		}
-
-		t.Log("CachePanicWithHandle stack handling test completed")
-	})
-
-	// 测试路径函数的错误返回路径
-	t.Run("path_functions_error_simulation", func(t *testing.T) {
-		// 测试ExecDir, ExecFile, Pwd函数的错误处理分支
-		// 在正常环境中这些函数很少失败，但我们需要确保错误处理代码被覆盖
-
-		// 调用所有路径函数确保它们被测试
-		functions := map[string]func() string{
-			"ExecDir":  ExecDir,
-			"ExecFile": ExecFile,
-			"Pwd":      Pwd,
-		}
-
-		for name, fn := range functions {
-			result := fn()
-			if result == "" {
-				t.Logf("%s returned empty - error path covered", name)
-			} else {
-				t.Logf("%s success path: %s", name, result)
-			}
-		}
-
-		t.Log("All path functions tested for error handling")
-	})
-
-	// 测试所有用户目录函数的完整逻辑
-	t.Run("user_directory_functions_complete", func(t *testing.T) {
-		// 确保所有用户目录函数的每一行都被执行
-		functions := map[string]func() string{
-			"UserHomeDir":   UserHomeDir,
-			"UserConfigDir": UserConfigDir,
-			"UserCacheDir":  UserCacheDir,
-			"LazyConfigDir": LazyConfigDir,
-			"LazyCacheDir":  LazyCacheDir,
-		}
-
-		for name, fn := range functions {
-			result := fn()
-			t.Logf("%s result: %s", name, result)
-		}
-
-		t.Log("All user directory functions fully tested")
-	})
-
-	// 测试所有系统检测函数
-	t.Run("system_detection_complete", func(t *testing.T) {
-		// 确保所有系统检测函数都被调用
-		results := map[string]bool{
-			"IsWindows": IsWindows(),
-			"IsDarwin":  IsDarwin(),
-			"IsLinux":   IsLinux(),
-		}
-
-		trueCount := 0
-		for name, result := range results {
-			if result {
-				trueCount++
-			}
-			t.Logf("%s: %v", name, result)
-		}
-
-		if trueCount != 1 {
-			t.Errorf("Expected exactly one OS to be true, got %d", trueCount)
-		}
-
-		t.Log("System detection functions fully tested")
-	})
-
-	// 测试信号相关函数（但避免实际信号发送）
-	t.Run("signal_functions_safe_test", func(t *testing.T) {
-		// 测试GetExitSign但不发送会干扰测试的信号
-		sigCh := GetExitSign()
-		if sigCh == nil {
-			t.Fatal("GetExitSign should not return nil")
-		}
-
-		// 验证通道容量
-		capacity := cap(sigCh)
-		t.Logf("Signal channel capacity: %d", capacity)
-
-		// 测试WaitExit的组件但不实际等待
-		// WaitExit只是调用GetExitSign然后从通道读取，我们已经测试了GetExitSign
-
-		t.Log("Signal functions tested safely")
-	})
-}
-
-// 测试所有panic处理函数的完整覆盖率
-func TestPanicHandlingFinal(t *testing.T) {
-	t.Run("cache_panic_all_paths", func(t *testing.T) {
-		// 测试CachePanic（它调用CachePanicWithHandle(nil)）
-		func() {
-			defer func() {
-				recover() // 防止panic传播
-			}()
-			defer CachePanic()
-			panic("test CachePanic")
-		}()
-
-		t.Log("CachePanic tested")
-	})
-
-	t.Run("cache_panic_with_handle_all_paths", func(t *testing.T) {
-		// 测试CachePanicWithHandle的所有路径
-
-		// 路径1: 无panic
-		func() {
-			defer CachePanicWithHandle(func(interface{}) {
-				t.Error("Should not be called when no panic")
-			})
-			// 正常执行
-		}()
-
-		// 路径2: 有panic，handle为nil
-		func() {
-			defer func() {
-				recover()
-			}()
-			defer CachePanicWithHandle(nil)
-			panic("test nil handle")
-		}()
-
-		// 路径3: 有panic，handle不为nil
-		handleCalled := false
-		func() {
-			defer func() {
-				recover()
-			}()
-			defer CachePanicWithHandle(func(err interface{}) {
-				handleCalled = true
-			})
-			panic("test with handle")
-		}()
-
-		if !handleCalled {
-			t.Error("Handle should have been called")
-		}
-
-		t.Log("All CachePanicWithHandle paths tested")
-	})
-}
-
-// 最终的覆盖率验证测试
-func TestCompleteSystemCoverage(t *testing.T) {
-	t.Run("verify_all_functions_called", func(t *testing.T) {
-		// 系统相关
-		_ = IsWindows()
-		_ = IsDarwin()
-		_ = IsLinux()
-
-		// 路径相关
-		_ = ExecDir()
-		_ = ExecFile()
-		_ = Pwd()
-
-		// 用户目录相关
-		_ = UserHomeDir()
-		_ = UserConfigDir()
-		_ = UserCacheDir()
-		_ = LazyConfigDir()
-		_ = LazyCacheDir()
-
-		// 信号相关
-		_ = GetExitSign()
-
-		// panic处理相关
-		PrintStack()
-		
-		func() {
-			defer CachePanic()
-		}()
-
-		func() {
-			defer CachePanicWithHandle(nil)
-		}()
-
-		func() {
-			defer CachePanicWithHandle(func(interface{}) {})
-		}()
-
-		t.Log("All functions verified for complete coverage")
-	})
-}
