--- runtime/missing_lines_test.go
@@ -1,127 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"os/exec"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-)
-
-// TestExitFunctionSpecific tests the Exit function that's currently uncovered
-func TestExitFunctionSpecific(t *testing.T) {
-	if os.Getenv("TEST_EXIT_SUBPROCESS") == "1" {
-		// This will be run in the subprocess to test Exit()
-		Exit() // This should trigger lines 20-32 in exit.go
-		return
-	}
-
-	// Run the test in a subprocess
-	cmd := exec.Command(os.Args[0], "-test.run=TestExitFunctionSpecific")
-	cmd.Env = append(os.Environ(), "TEST_EXIT_SUBPROCESS=1")
-	
-	err := cmd.Start()
-	assert.NoError(t, err)
-
-	// Give the process time to run Exit()
-	done := make(chan error, 1)
-	go func() {
-		done <- cmd.Wait()
-	}()
-
-	select {
-	case <-done:
-		// Process exited/terminated - this is expected behavior
-		t.Log("Exit function test completed - subprocess terminated as expected")
-	case <-time.After(3 * time.Second):
-		// Kill if it hangs
-		if cmd.Process != nil {
-			cmd.Process.Kill()
-		}
-		t.Log("Subprocess terminated after timeout")
-	}
-}
-
-// TestEmptyStackCondition attempts to test the "else" branch in CachePanicWithHandle (lines 26-27)
-// and PrintStack (lines 41-42). These are very difficult to trigger in normal Go runtime.
-func TestEmptyStackCondition(t *testing.T) {
-	// The empty stack condition is nearly impossible to trigger in normal circumstances
-	// because debug.Stack() almost always returns a non-empty stack trace in Go.
-	// However, we can test the normal code path to ensure these functions work correctly.
-	
-	t.Run("cache_panic_with_stack", func(t *testing.T) {
-		panicHandled := false
-		
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Test recovered panic: %v", r)
-				}
-			}()
-			
-			func() {
-				defer CachePanicWithHandle(func(err interface{}) {
-					panicHandled = true
-					assert.Equal(t, "stack test", err)
-				})
-				panic("stack test")
-			}()
-		}()
-		
-		assert.True(t, panicHandled)
-	})
-	
-	t.Run("print_stack_normal", func(t *testing.T) {
-		// PrintStack should work normally
-		PrintStack() // This tests the normal path (lines 38-40)
-	})
-}
-
-// TestErrorPathsFilesystem tests error conditions in filesystem functions
-// These error paths (lines 48-49, 56-57, 64-65) are hard to trigger but we'll try
-func TestErrorPathsFilesystem(t *testing.T) {
-	// The error paths in ExecDir, ExecFile, and Pwd are triggered when:
-	// - os.Executable() fails
-	// - os.Getwd() fails
-	// These are very rare in normal conditions, but let's test the normal paths
-	
-	t.Run("exec_dir_coverage", func(t *testing.T) {
-		dir := ExecDir()
-		// In normal conditions, this should return a valid directory
-		if dir == "" {
-			// This would hit the error path (line 49-50)
-			t.Log("ExecDir returned empty (error path)")
-		} else {
-			// This is the normal path (line 51)
-			assert.NotEmpty(t, dir)
-			t.Logf("ExecDir: %s", dir)
-		}
-	})
-	
-	t.Run("exec_file_coverage", func(t *testing.T) {
-		file := ExecFile()
-		// In normal conditions, this should return a valid file path
-		if file == "" {
-			// This would hit the error path (line 57-58)
-			t.Log("ExecFile returned empty (error path)")
-		} else {
-			// This is the normal path (line 59)
-			assert.NotEmpty(t, file)
-			t.Logf("ExecFile: %s", file)
-		}
-	})
-	
-	t.Run("pwd_coverage", func(t *testing.T) {
-		pwd := Pwd()
-		// In normal conditions, this should return a valid directory
-		if pwd == "" {
-			// This would hit the error path (line 65-66)
-			t.Log("Pwd returned empty (error path)")
-		} else {
-			// This is the normal path (line 67)
-			assert.NotEmpty(t, pwd)
-			t.Logf("Pwd: %s", pwd)
-		}
-	})
-}
