--- runtime/mock_error_test.go
@@ -1,185 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"testing"
-)
-
-// 尝试通过模拟错误情况来提高覆盖率
-func TestMockErrorConditions(t *testing.T) {
-	t.Run("simulate_error_conditions", func(t *testing.T) {
-		// 由于很难真正触发os.Executable(), os.Getwd()等函数的错误
-		// 我们主要测试这些函数的正常路径，以确保它们被执行
-
-		// 测试ExecDir - 正常情况下应该成功
-		execDir := ExecDir()
-		if execDir == "" {
-			t.Log("ExecDir returned empty - error branch covered")
-		} else {
-			t.Logf("ExecDir returned: %s - success branch covered", execDir)
-		}
-
-		// 测试ExecFile - 正常情况下应该成功
-		execFile := ExecFile()
-		if execFile == "" {
-			t.Log("ExecFile returned empty - error branch covered")
-		} else {
-			t.Logf("ExecFile returned: %s - success branch covered", execFile)
-		}
-
-		// 测试Pwd - 正常情况下应该成功
-		pwd := Pwd()
-		if pwd == "" {
-			t.Log("Pwd returned empty - error branch covered")
-		} else {
-			t.Logf("Pwd returned: %s - success branch covered", pwd)
-		}
-	})
-
-	t.Run("test_print_stack_empty_stack_branch", func(t *testing.T) {
-		// 测试PrintStack函数 - 正常情况下debug.Stack()不会返回空
-		// 但我们确保函数被完整执行
-
-		PrintStack()
-		t.Log("PrintStack executed - should have valid stack trace")
-	})
-
-	t.Run("test_cache_panic_with_handle_complete_coverage", func(t *testing.T) {
-		// 确保CachePanicWithHandle的所有分支都被覆盖
-
-		// 测试1: 没有panic的情况
-		func() {
-			defer CachePanicWithHandle(func(err interface{}) {
-				t.Error("Handle should not be called when no panic")
-			})
-			// 正常执行，不panic
-		}()
-
-		// 测试2: 有panic且handle不为nil
-		handleCalled := false
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Expected panic recovered: %v", r)
-				}
-				if !handleCalled {
-					t.Error("Handle should have been called")
-				}
-			}()
-
-			defer CachePanicWithHandle(func(err interface{}) {
-				handleCalled = true
-				if err != "test panic" {
-					t.Errorf("Wrong error in handle: %v", err)
-				}
-			})
-			panic("test panic")
-		}()
-
-		// 测试3: 有panic且handle为nil
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					t.Logf("Expected panic with nil handle recovered: %v", r)
-				}
-			}()
-
-			defer CachePanicWithHandle(nil)
-			panic("test panic with nil handle")
-		}()
-
-		t.Log("CachePanicWithHandle complete coverage test finished")
-	})
-}
-
-// 测试Exit函数的相关组件（但不调用Exit本身）
-func TestExitRelatedComponents(t *testing.T) {
-	t.Run("test_exit_signal_setup", func(t *testing.T) {
-		// 测试GetExitSign函数的完整功能
-		sigCh := GetExitSign()
-
-		if sigCh == nil {
-			t.Fatal("GetExitSign should not return nil")
-		}
-
-		// 验证通道属性
-		select {
-		case sigCh <- os.Interrupt:
-			// 通道可以接收信号，说明是缓冲的
-			t.Log("Signal channel is properly buffered")
-		default:
-			t.Error("Signal channel should be buffered")
-		}
-
-		// 清理通道
-		select {
-		case <-sigCh:
-		default:
-		}
-	})
-}
-
-// 尝试提高函数覆盖率的综合测试
-func TestComprehensiveCoverage(t *testing.T) {
-	t.Run("ensure_all_functions_called", func(t *testing.T) {
-		// 确保所有主要函数都被调用，以最大化覆盖率
-
-		// 信号相关函数
-		sigCh := GetExitSign()
-		if sigCh != nil {
-			t.Log("GetExitSign: ✓")
-		}
-
-		// Panic处理函数
-		CachePanic() // 无panic情况
-		t.Log("CachePanic: ✓")
-
-		CachePanicWithHandle(nil) // 无panic情况
-		t.Log("CachePanicWithHandle: ✓")
-
-		// 堆栈函数
-		PrintStack()
-		t.Log("PrintStack: ✓")
-
-		// 路径函数
-		if ExecDir() != "" || true {
-			t.Log("ExecDir: ✓")
-		}
-
-		if ExecFile() != "" || true {
-			t.Log("ExecFile: ✓")
-		}
-
-		if Pwd() != "" || true {
-			t.Log("Pwd: ✓")
-		}
-
-		// 用户目录函数
-		UserHomeDir()
-		t.Log("UserHomeDir: ✓")
-
-		UserConfigDir()
-		t.Log("UserConfigDir: ✓")
-
-		UserCacheDir()
-		t.Log("UserCacheDir: ✓")
-
-		LazyConfigDir()
-		t.Log("LazyConfigDir: ✓")
-
-		LazyCacheDir()
-		t.Log("LazyCacheDir: ✓")
-
-		// 系统检测函数
-		IsWindows()
-		t.Log("IsWindows: ✓")
-
-		IsDarwin()
-		t.Log("IsDarwin: ✓")
-
-		IsLinux()
-		t.Log("IsLinux: ✓")
-
-		t.Log("All testable functions have been called")
-	})
-}
