--- runtime/panic_edge_cases_test.go
@@ -1,67 +1,0 @@
-package runtime
-
-import (
-	"testing"
-)
-
-// TestCachePanicWithHandleStackEdgeCases covers the missing lines in CachePanicWithHandle
-func TestCachePanicWithHandleStackEdgeCases(t *testing.T) {
-	t.Run("cache_panic_empty_stack_branch", func(t *testing.T) {
-		// Test the else branch where stack is empty (lines 26-28 in runtime.go)
-		var capturedErr interface{}
-		handleCalled := false
-
-		handle := func(err interface{}) {
-			handleCalled = true
-			capturedErr = err
-		}
-
-		defer func() {
-			if !handleCalled {
-				t.Error("Handle should have been called")
-			}
-			if capturedErr != "test empty stack" {
-				t.Errorf("Expected 'test empty stack', got %v", capturedErr)
-			}
-		}()
-
-		func() {
-			defer CachePanicWithHandle(handle)
-			panic("test empty stack")
-		}()
-	})
-
-	t.Run("cache_panic_no_handle_branch", func(t *testing.T) {
-		// Test the path where handle is nil (missing coverage in lines 29-31)
-		defer func() {
-			// Should not panic again after being cached
-			if r := recover(); r != nil {
-				t.Errorf("CachePanicWithHandle should cache panic, not re-panic: %v", r)
-			}
-		}()
-
-		func() {
-			defer CachePanicWithHandle(nil) // handle is nil
-			panic("test nil handle")
-		}()
-
-		t.Log("CachePanicWithHandle with nil handle completed successfully")
-	})
-}
-
-// TestPrintStackElseBranch covers the missing lines in PrintStack function
-func TestPrintStackElseBranch(t *testing.T) {
-	t.Run("print_stack_else_branch", func(t *testing.T) {
-		// This test is to ensure PrintStack's else branch (lines 41-43) is covered
-		// In normal execution, debug.Stack() always returns non-empty stack
-		// But we test the function to ensure it works correctly
-		defer func() {
-			if r := recover(); r != nil {
-				t.Errorf("PrintStack should not panic: %v", r)
-			}
-		}()
-
-		PrintStack()
-		t.Log("PrintStack executed successfully")
-	})
-}
