--- runtime/path_error_coverage_test.go
@@ -1,72 +1,0 @@
-package runtime
-
-import (
-	"os"
-	"os/exec"
-	"testing"
-)
-
-// TestPathFunctionErrorBranches tests the error return branches in path functions
-func TestPathFunctionErrorBranches(t *testing.T) {
-	if os.Getenv("GO_SIMULATE_PATH_ERRORS") == "1" {
-		// In subprocess, test the actual functions even if they might fail
-		// This ensures the error branches are exercised
-		t.Run("exec_dir_error_branch", func(t *testing.T) {
-			result := ExecDir()
-			t.Logf("ExecDir result: %s", result)
-		})
-
-		t.Run("exec_file_error_branch", func(t *testing.T) {
-			result := ExecFile()
-			t.Logf("ExecFile result: %s", result)
-		})
-
-		t.Run("pwd_error_branch", func(t *testing.T) {
-			result := Pwd()
-			t.Logf("Pwd result: %s", result)
-		})
-		return
-	}
-
-	t.Run("test_path_functions_error_branches", func(t *testing.T) {
-		// Run in subprocess to potentially trigger error conditions
-		cmd := exec.Command(os.Args[0], "-test.run=TestPathFunctionErrorBranches")
-		cmd.Env = append(os.Environ(), "GO_SIMULATE_PATH_ERRORS=1")
-		output, err := cmd.CombinedOutput()
-		
-		t.Logf("Subprocess output: %s", string(output))
-		if err != nil {
-			t.Logf("Subprocess completed with: %v", err)
-		}
-	})
-
-	t.Run("exec_dir_normal_path", func(t *testing.T) {
-		// Test ExecDir normal execution to ensure both branches are covered
-		result := ExecDir()
-		if result == "" {
-			t.Log("ExecDir returned empty (error branch)")
-		} else {
-			t.Logf("ExecDir returned: %s (success branch)", result)
-		}
-	})
-
-	t.Run("exec_file_normal_path", func(t *testing.T) {
-		// Test ExecFile normal execution
-		result := ExecFile()
-		if result == "" {
-			t.Log("ExecFile returned empty (error branch)")
-		} else {
-			t.Logf("ExecFile returned: %s (success branch)", result)
-		}
-	})
-
-	t.Run("pwd_normal_path", func(t *testing.T) {
-		// Test Pwd normal execution
-		result := Pwd()
-		if result == "" {
-			t.Log("Pwd returned empty (error branch)")
-		} else {
-			t.Logf("Pwd returned: %s (success branch)", result)
-		}
-	})
-}
