--- runtime/stack_edge_test.go
@@ -1,239 +1,0 @@
-package runtime
-
-import (
-	"runtime/debug"
-	"testing"
-)
-
-// 测试PrintStack函数的边界情况
-func TestPrintStackEdgeCases(t *testing.T) {
-	t.Run("test_print_stack_normal_execution", func(t *testing.T) {
-		// 测试PrintStack在正常情况下的执行
-		// 这应该触发有堆栈信息的分支
-		defer func() {
-			if r := recover(); r != nil {
-				t.Errorf("PrintStack should not panic: %v", r)
-			}
-		}()
-
-		PrintStack()
-		t.Log("PrintStack executed successfully with stack information")
-	})
-
-	t.Run("test_print_stack_function_coverage", func(t *testing.T) {
-		// 确保PrintStack函数的所有代码路径都被执行
-		// 在正常情况下，debug.Stack()总是返回非空的堆栈信息
-		stack := debug.Stack()
-		if len(stack) == 0 {
-			t.Log("debug.Stack() returned empty - rare case detected")
-		} else {
-			t.Logf("debug.Stack() returned %d bytes of stack information", len(stack))
-		}
-
-		// 调用PrintStack来确保覆盖
-		PrintStack()
-		t.Log("PrintStack function call completed")
-	})
-
-	t.Run("test_print_stack_log_branches", func(t *testing.T) {
-		// 测试PrintStack函数中的日志分支
-		// 由于debug.Stack()在正常情况下不会返回空，
-		// 我们主要确保函数被完整执行
-		
-		// 多次调用以确保稳定性
-		for i := 0; i < 3; i++ {
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Errorf("PrintStack call %d should not panic: %v", i, r)
-					}
-				}()
-				PrintStack()
-			}()
-		}
-		
-		t.Log("Multiple PrintStack calls completed successfully")
-	})
-}
-
-// 测试CachePanicWithHandle函数的堆栈处理边界情况
-func TestCachePanicStackHandling(t *testing.T) {
-	t.Run("test_cache_panic_stack_normal", func(t *testing.T) {
-		// 测试CachePanicWithHandle在正常panic情况下的堆栈处理
-		handleCalled := false
-		var capturedErr interface{}
-
-		handle := func(err interface{}) {
-			handleCalled = true
-			capturedErr = err
-		}
-
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Test recovered panic: %v", r)
-			}
-			
-			if !handleCalled {
-				t.Error("Handle should have been called")
-			}
-			
-			if capturedErr == nil {
-				t.Error("Handle should have received error")
-			}
-		}()
-
-		func() {
-			defer CachePanicWithHandle(handle)
-			panic("test stack handling")
-		}()
-	})
-
-	t.Run("test_cache_panic_stack_logging", func(t *testing.T) {
-		// 测试CachePanicWithHandle中堆栈日志记录的完整路径
-		// 确保所有日志分支都被执行
-		
-		for i := 0; i < 2; i++ {
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Logf("Iteration %d: panic recovered: %v", i, r)
-					}
-				}()
-
-				defer CachePanicWithHandle(func(err interface{}) {
-					t.Logf("Iteration %d: handle called with error: %v", i, err)
-				})
-				
-				panic("test iteration " + string(rune('0'+i)))
-			}()
-		}
-		
-		t.Log("Stack logging test completed")
-	})
-
-	t.Run("test_cache_panic_empty_stack_simulation", func(t *testing.T) {
-		// 虽然我们不能真正模拟debug.Stack()返回空，
-		// 但我们可以确保CachePanicWithHandle的所有分支都被测试
-		
-		// 测试不同类型的panic值
-		panicValues := []interface{}{
-			"string panic",
-			42,
-			struct{ msg string }{"struct panic"},
-			nil,
-		}
-
-		for i, panicValue := range panicValues {
-			func() {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Logf("Panic value %d recovered: %v", i, r)
-					}
-				}()
-
-				defer CachePanicWithHandle(func(err interface{}) {
-					t.Logf("Handle for panic value %d: %v", i, err)
-				})
-				
-				panic(panicValue)
-			}()
-		}
-		
-		t.Log("Different panic value types tested")
-	})
-}
-
-// 测试运行时函数的极端情况
-func TestRuntimeFunctionExtreme(t *testing.T) {
-	t.Run("test_deep_call_stack", func(t *testing.T) {
-		// 创建深度调用栈来测试PrintStack
-		var deepCall func(int)
-		deepCall = func(depth int) {
-			if depth <= 0 {
-				PrintStack()
-				return
-			}
-			deepCall(depth - 1)
-		}
-
-		defer func() {
-			if r := recover(); r != nil {
-				t.Errorf("Deep call stack should not cause panic: %v", r)
-			}
-		}()
-
-		deepCall(10) // 创建10层深的调用栈
-		t.Log("Deep call stack PrintStack test completed")
-	})
-
-	t.Run("test_concurrent_print_stack", func(t *testing.T) {
-		// 测试并发调用PrintStack的安全性
-		const numGoroutines = 5
-		done := make(chan bool, numGoroutines)
-
-		for i := 0; i < numGoroutines; i++ {
-			go func(id int) {
-				defer func() {
-					if r := recover(); r != nil {
-						t.Errorf("Goroutine %d: PrintStack should not panic: %v", id, r)
-					}
-					done <- true
-				}()
-
-				PrintStack()
-			}(i)
-		}
-
-		// 等待所有goroutine完成
-		for i := 0; i < numGoroutines; i++ {
-			<-done
-		}
-
-		t.Log("Concurrent PrintStack test completed")
-	})
-}
-
-// 测试panic处理的完整覆盖率
-func TestPanicHandlingComplete(t *testing.T) {
-	t.Run("test_cache_panic_function_complete", func(t *testing.T) {
-		// 测试CachePanic函数（它调用CachePanicWithHandle(nil)）
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("CachePanic test recovered: %v", r)
-			}
-		}()
-
-		func() {
-			defer CachePanic()
-			panic("test CachePanic function")
-		}()
-
-		t.Log("CachePanic function test completed")
-	})
-
-	t.Run("test_nested_panic_handling", func(t *testing.T) {
-		// 测试嵌套的panic处理
-		defer func() {
-			if r := recover(); r != nil {
-				t.Logf("Outer recover: %v", r)
-			}
-		}()
-
-		func() {
-			defer func() {
-				defer CachePanicWithHandle(func(err interface{}) {
-					t.Logf("Inner handle: %v", err)
-				})
-				panic("inner panic")
-			}()
-			
-			defer CachePanicWithHandle(func(err interface{}) {
-				t.Logf("Outer handle: %v", err)
-			})
-			
-			panic("outer panic")
-		}()
-
-		t.Log("Nested panic handling test completed")
-	})
-}
